+++ src/main/java/com/capgemini/gradebook/service/impl/StudentServiceImpl.java
@@ -19,6 +19,7 @@ import com.capgemini.gradebook.persistence.repo.ClassRepo;
 import com.capgemini.gradebook.persistence.repo.StudentRepo;
 import com.capgemini.gradebook.service.StudentService;
 import com.capgemini.gradebook.service.exception.BadRequestException;
+import com.capgemini.gradebook.service.exception.NotFoundException;
 
 @Service
 @Transactional
@@ -42,67 +43,68 @@ public class StudentServiceImpl implements StudentService {
   @Override
   public StudentEto findStudentById(Long id) {
     final Optional<StudentEntity> result = this.studentRepository.findById(id);
-    return result.map(r -> StudentMapper.mapToETO(r)).orElseThrow( () -> new RuntimeException("Student not found " +
-        "exception"));
+    return result.map(r -> StudentMapper.mapToETO(r)).orElseThrow( () -> new NotFoundException("Student not found"));
   }
 
   @Override
   public StudentEto save(StudentEto newStudent) {
 
     if (newStudent.getId() != null) {
-      throw new BadRequestException("Teacher with such id already exists");
+      throw new BadRequestException("Student with such id already exists");
     }
 
     StudentEntity studentEntity = StudentMapper.mapToEntity(newStudent);
     if(newStudent.getClassYearId() == null) {
-      throw new RuntimeException("student has no class year");
+      throw new NotFoundException("Student has no class year");
     } else {
       Optional<ClassYearEntity> classYearEntity = this.classYearRepository.findById(newStudent.getClassYearId());
       if(classYearEntity.isPresent()) {
         studentEntity.setClassYearEntity(classYearEntity.get());
       } else {
-        throw new RuntimeException("There is no class year with such id in db");
+        throw new NotFoundException("There is no class year with such id in db");
       }
     }
     studentEntity = this.studentRepository.save(studentEntity);
     return StudentMapper.mapToETO(studentEntity);
   }
 
-//  @Override
-//  public StudentEto updateStudent(StudentEto studentEto) {
-//    String studentFirstName = studentEto.getFirstName();
-//    String studentLastName = studentEto.getLastName();
-//    int studentAge = studentEto.getAge();
-//    Long studentId = studentEto.getId();
-//    ClassYearEntity classYear = classYearRepository.findById(studentEto.getClassYearId()).get();
-//
-//    StudentEntity student =
-//        this.studentRepository.findById(studentId).orElseThrow(() -> new IllegalStateException(
-//            "Student with" +
-//                " id " + studentEto.getId() + " doesn't exist"));
-//
-//    if(studentFirstName != null && studentFirstName.length() > 0 && !Objects.equals(student.getFirstName(),
-//        studentFirstName)) {
-//      student.setFirstName(studentFirstName);
-//    }
-//
-//    if(studentLastName != null && studentLastName.length() > 0 && !Objects.equals(student.getLastName(),
-//        studentLastName)) {
-//      student.setLastName(studentLastName);
-//    }
-//
-//    if(studentAge > 0 && !Objects.equals(student.getAge(), studentAge)) {
-//      student.setAge(studentAge);
-//    }
-//
-//    if(classYear != null && > 0 && !Objects.equals(student.getAge(), studentAge)) {
-//      student.setAge(studentAge);
-//    }
-//    return StudentMapper.mapToETO(student);
-//  }
+  @Override
+  public StudentEto updateStudent(StudentEto studentEto, Long studentId) {
+    String studentFirstName = studentEto.getFirstName();
+    String studentLastName = studentEto.getLastName();
+    int studentAge = studentEto.getAge();
+    ClassYearEntity classYear = classYearRepository.findById(studentEto.getClassYearId()).get();
+
+    StudentEntity student =
+        this.studentRepository.findById(studentId).orElseThrow(() -> new IllegalStateException(
+            "Student with id " + studentEto.getId() + " doesn't exist"));
+
+    if(studentFirstName != null && studentFirstName.length() > 0 && !Objects.equals(student.getFirstName(),
+        studentFirstName)) {
+      student.setFirstName(studentFirstName);
+    }
+
+    if(studentLastName != null && studentLastName.length() > 0 && !Objects.equals(student.getLastName(),
+        studentLastName)) {
+      student.setLastName(studentLastName);
+    }
+
+    if(studentAge > 0 && !Objects.equals(student.getAge(), studentAge)) {
+      student.setAge(studentAge);
+    }
+
+    if(classYear != null && !Objects.equals(student.getClassYearEntity(), classYear)) {
+      student.setClassYearEntity(classYear);
+    }
+    return StudentMapper.mapToETO(student);
+  }
 
   @Override
   public void delete(Long id) {
+    if (!studentRepository.existsById(id)) {
+      throw new NotFoundException("Student with id " + id + " does not exist");
+    }
+
     this.studentRepository.deleteById(id);
   }
 }
